entity-linking/requirements.txt
python-dotenv==0.21.1
elasticsearch==8.6.1
python-Levenshtein==0.20.9
tqdm==4.64.1
beautifulsoup4==4.11.2
selenium==4.8.0
requests==2.28.2
sparqlwrapper==2.0.0
matplotlib==3.7.0
seaborn==0.12.2